% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_formatting.R
\name{add_period}
\alias{add_period}
\title{FUNCTION_TITLE}
\usage{
add_period(str)
}
\arguments{
\item{str}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
