% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_conversions.R
\name{base2type}
\alias{base2type}
\title{FUNCTION_TITLE}
\usage{
base2type(base)
}
\arguments{
\item{base}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
