% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_assignments.R
\name{build_hw_assignment}
\alias{build_hw_assignment}
\title{Build a homework assignment}
\usage{
build_hw_assignment(schedule, date, cal_entry, semester, dry_run = FALSE)
}
\arguments{
\item{schedule}{A schedule data frame.}

\item{date}{The homework due date.}

\item{cal_entry}{A calendar entry for the homework due date.}

\item{semester}{A list of data for the semester, from the database.}

\item{dry_run}{Don't actually write assignment files to disk.}
}
\value{
An updated schedule data frame
}
\description{
Build a homework assignment: Generate an \code{.Rmd} file for a
homework assignment web page and PDF handout.
}
