% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pdf_files.R
\name{build_pdf_files}
\alias{build_pdf_files}
\title{Build PDF versions of assignments.}
\usage{
build_pdf_files(
  semester,
  content_path = "content",
  static_path = "static",
  force_dest = TRUE,
  output_options = NULL
)
}
\arguments{
\item{semester}{A list of data returned from \verb{\link{load_semester_db}}}

\item{content_path}{A relative path to the HUGO content directory.}

\item{static_path}{Relative path to the HUGO static directory.}

\item{force_dest}{Create any missing directories.}

\item{output_options}{A list of extra output options
for \verb{\link{pdf_document}}.}
}
\value{
A vector of paths to the resulting PDF files.
}
\description{
Build PDF versions of assignments RMarkdown.
}
