% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pdf_files.R
\name{build_pdf_output_format}
\alias{build_pdf_output_format}
\title{Build an RMarkdown output format for PDF files.}
\usage{
build_pdf_output_format(header, output_options = NULL)
}
\arguments{
\item{header}{Metadata from the RMarkdown file.}

\item{output_options}{A list of extra output options
for \verb{\link{pdf_document}}.}
}
\value{
An RMarkdown output format object.
}
\description{
Build an RMarkdown output format for PDF files.
}
