% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_assignments.R
\name{build_reading_assignment}
\alias{build_reading_assignment}
\title{Build a reading assignment}
\usage{
build_reading_assignment(schedule, date, cal_entry, semester, dry_run = FALSE)
}
\arguments{
\item{schedule}{A schedule data frame.}

\item{date}{The reading due date.}

\item{cal_entry}{A calendar entry for the class.}

\item{semester}{A list of data for the semester, from the database.}

\item{dry_run}{Don't actually write assignment files to disk.}
}
\value{
An updated schedule data frame
}
\description{
Build a reading assignment: Generate an \code{.Rmd} file for a
reading assignment web page and PDF handout.
}
