% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_database.R
\name{cal_prepare}
\alias{cal_prepare}
\title{Convert an event-specific calendar tibble to a general master calendar.}
\usage{
cal_prepare(df, type, key_col = NULL)
}
\arguments{
\item{df}{A dataframe holding the calendar. Columns should include:
\describe{
\item{date}{The date of the calendar item}
\item{\code{key_col}}{A character string with the key for the calendar entry.
This column's name should match the \code{key_col} argument.}
\item{cancelled}{An integer indicating whether the entry has been
cancelled for the date of this entry.}
\item{make_up}{An integer indicating whether the entry represents a
re-scheduled make-up event}
}}

\item{type}{The type of event. This should be one of \code{class}, \code{lab},
\code{homework}, \verb{due date}, \code{exam}, \code{holiday}, or \code{event}.
If \code{type} is \code{NULL}, the value is constructed by removing "\verb{_key}" from
the end of the \code{key_col} argument (e.g., "\code{class_key}" becomes "\code{class}").}

\item{key_col}{The name of the column with the keys for the entries to be
dplyr::selected. This can either be quoted or unquoted (i.e., using non-standard
tidy evaluation). If the \code{key_col} argument is missing and \code{type} is not
\code{NULL}, then a default \code{key_col} will be constructed by appending "\verb{_key}"
to \code{type} (e.g., "\code{class}" becomes "\code{class_key}").}
}
\value{
A tibble with columns:
\describe{
\item{cal_id}{An integer whose thousands place indicates the kind of
entry, hundreds place indicates re-scheduling (cancellation or make-up),
and last two digits indicate the order of the entry.}
\item{date}{The date of the entry, as an ISO format string.}
\item{cancelled}{An integer indicating whether the entry has been
cancelled.}
\item{make_up}{An integer indicating whether the entry represents a
re-scheduled make-up session.}
\item{cal_key}{A column with a text key for that entry. }
\item{cal_type}{A string indicating the type of entry (class, lab, etc.)}
\item{...}{Other entries, such as a numerical index for events of a
given type. In general, we expect (but do not require) that there is
a numerical index column corresponding to the \code{key_col} argument
(e.g., if \code{key_col} = "\code{lab_key}", then there should be a "\code{lab_num}"
column corresponding to the sequence of laboratory entries).}
}
}
\description{
Convert the columns of a calendar specific to a certain kind of entry to
generic columns suitable for a master calendar containing multiple kinds of
entries.
}
