% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_assignments.R
\name{check_schedule}
\alias{check_schedule}
\title{Check a schedule data frame for consistency}
\usage{
check_schedule(schedule, semester)
}
\arguments{
\item{schedule}{A schedule data frame.}

\item{semester}{A list of data for the semester, from the database.}
}
\value{
nothing
}
\description{
Runs some basic consistency checks on a schedule data frame
}
