% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_assignments.R
\name{clean_assignment_files}
\alias{clean_assignment_files}
\title{Clean up assignment riles.}
\usage{
clean_assignment_files(verbose = TRUE, dry_run = TRUE)
}
\arguments{
\item{verbose}{Send progress messages to the console}

\item{dry_run}{Simulate the cleaning without actually deleting files.}
}
\description{
Clean up assignment files to prepare for freshly regenerating assignments.
}
