% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_conversions.R
\name{col2type}
\alias{col2type}
\title{FUNCTION_TITLE}
\usage{
col2type(col)
}
\arguments{
\item{col}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
