% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_search.R
\name{col_replace}
\alias{col_replace}
\title{Search and replace all columns in a data frame.}
\usage{
col_replace(df, pattern, replacement)
}
\arguments{
\item{df}{A data frame}

\item{pattern}{The pattern to search for (regular expression)}

\item{replacement}{The replacement.}
}
\value{
A data frame with the pattern replaced in every column.
}
\description{
Search and replace in every column of a data frame.
}
