% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_search.R
\name{col_search}
\alias{col_search}
\title{Search for a pattern in any column of a data frame}
\usage{
col_search(df, pattern)
}
\arguments{
\item{df}{A data frame}

\item{pattern}{The pattern to search for (regular expression)}
}
\value{
A data frame where each character column is replaced with "HIT" if
it matches the pattern and "" if it doesn't.
}
\description{
Search every character column of a data frame for a pattern.
}
