% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_formatting.R
\name{concat_with_nl}
\alias{concat_with_nl}
\title{Concatenate strings as separate lines of text.}
\usage{
concat_with_nl(s, ..., start_par = FALSE, extra_lines = 0, collapse = "")
}
\arguments{
\item{s}{A character string}

\item{...}{Additional character strings}

\item{start_par}{Put an extra newline between each pair of strings, so each
string starts a new paragraph.}

\item{extra_lines}{Number of extra newlines to append after each string.}

\item{collapse}{Extra characters to insert between the strings.}
}
\value{
A character string.
}
\description{
Combine strings, as separate lines of text.
}
\details{
Take two or more lines of text. Make sure they each end with a newline and
then concatenate them.
}
\examples{
concat_with_nl("foo", "bar")
concat_with_nl("foo", "bar", start_par = TRUE)
concat_with_nl("foo", "bar", collapse = "...")

}
