% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_formatting.R
\name{date_parts}
\alias{date_parts}
\alias{format_month}
\alias{format_wday}
\title{Formatting parts of dates}
\usage{
format_month(d, abbr = TRUE)

format_wday(d, abbr = TRUE)
}
\arguments{
\item{d}{A date}

\item{abbr}{Abbreviate the name.}
}
\value{
The name of the month or day of the week.
}
\description{
Format the month or day of week for a date.
}
\section{Functions}{
\itemize{
\item \code{format_month}: Format the month

\item \code{format_wday}: Format the day of the week
}}

\examples{
format_month(ymd("2001-02-25"), FALSE)
format_month(lubridate::now(), TRUE)

format_wday(ymd("2001-02-25"), FALSE)
format_wday(Sys.Date(), TRUE)

}
