% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_utils.R
\name{default_semestr_metadata}
\alias{default_semestr_metadata}
\title{FUNCTION_TITLE}
\usage{
default_semestr_metadata()
}
\arguments{
\item{DESCRIPTION.}{}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
