% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_database.R
\name{do_index}
\alias{do_index}
\title{Extract and index calendar entries by type.}
\usage{
do_index(df, key_col, type = NULL)
}
\arguments{
\item{df}{A dataframe holding the calendar. Columns should include:
\describe{
\item{date}{The date of the calendar item}
\item{\code{key_col}}{A character string with the key for the calendar entry.
This column's name should match the \code{key_col} argument.}
\item{cancelled}{An integer indicating whether the entry has been
cancelled for the date of this entry.}
\item{make_up}{An integer indicating whether the entry represents a
re-scheduled make-up event}
}}

\item{key_col}{The name of the column with the keys for the entries to be
selected. This can either be quoted or unquoted (i.e., using non-standard
tidy evaluation). If the \code{key_col} argument is missing and \code{type} is not
\code{NULL}, then a default \code{key_col} will be constructed by appending "\verb{_key}"
to \code{type} (e.g., "\code{class}" becomes "\code{class_key}").}

\item{type}{The type of event. This should be one of \code{class}, \code{lab},
\code{homework}, \verb{due date}, \code{exam}, \code{holiday}, or \code{event}.
If \code{type} is \code{NULL}, the value is constructed by removing "\verb{_key}" from
the end of the \code{key_col} argument (e.g., "\code{class_key}" becomes "\code{class}").}
}
\value{
A tibble with columns:
\describe{
\item{cal_id}{An integer whose thousands place indicates the kind of
entry, hundreds place indicates re-scheduling (cancellation or make-up),
and last two digits indicate the order of the entry.}
\item{date}{The date of the entry, as an ISO format string.}
\item{cancelled}{An integer indicating whether the entry has been
cancelled.}
\item{make_up}{An integer indicating whether the entry represents a
re-scheduled make-up session.}
\item{\verb{_key}}{A column with a text key for that entry. The name of this
column is the \code{key_col} argument.}
\item{\verb{_num}}{A column with an integer index representing the order of the
entry in the sequence of this kind of entry (class, lab, etc.). The
name of this column is constructed by replacing "\verb{_key}" with "\verb{_num}"
in the \code{key_col} argument (e.g., "\code{class_key}" becomes "\code{class_num}").}
\item{cal_type}{A string indicating the type of entry (class, lab, etc.)}
}
}
\description{
Extract calendar entries of a given type. Add an index indicating the
sequence of entries with that type, and add a column indicating what
the type is.
}
\details{
The type is based on the \code{key_col} argument: Entries with a key in that
column are chosen, and the numerical index column is created using a name
derived by substituting "\verb{_num}" ror "\verb{_key}" in the column name (e.g.,
"\code{class_key}" becomes "\code{class_num}"). The index represents the sequence of
events of the chosen type, ordered by date.
}
