% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_formatting.R
\name{escape_dollar}
\alias{escape_dollar}
\title{Escape dollar sign for LaTeX}
\usage{
escape_dollar(txt)
}
\arguments{
\item{txt}{A character string}
}
\value{
A character string with escaped backslashes in front of dollar sings.
}
\description{
Escape single backslashes in front of dollar signs (for LaTeX)
}
\details{
If a dollar sign in a character string is preceded by a single backslash,
escape that to make it a double-backslash. Ignore cases where the dollar
sign is already preceded by multiple backslashes.
}
\examples{
escape_dollar("$x^2$")
escape_dollar("\\\\$1000.00")
escape_dollar("$\\\\\\\\$1000.00")

}
