% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_semester_db.R
\name{fixup_semester}
\alias{fixup_semester}
\title{Fix up semester object for canceled and makeup classes}
\usage{
fixup_semester(semester)
}
\arguments{
\item{semester}{A semester object returned by \code{load_semester_db()}.}
}
\value{
A fixed up \code{semester} object.
}
\description{
Fix up calendar and rd_items to account for classes that are
canceled and rescheduled for makeup.
}
\examples{
\dontrun{
semester <- fixup_semester(semester)
}
}
