% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_formatting.R
\name{format_class_date}
\alias{format_class_date}
\alias{format_class_day_date}
\title{Format a date}
\usage{
format_class_date(d, abbr = TRUE)

format_class_day_date(d, abbr_month = TRUE, abbr_wday = TRUE)
}
\arguments{
\item{d}{A date.}

\item{abbr}{Abbreviate the month.}

\item{abbr_month}{Abbreviate the month.}

\item{abbr_wday}{Abbreviate the day of the week.}
}
\value{
The formatted date.
}
\description{
Format a date as month and day.
}
\section{Functions}{
\itemize{
\item \code{format_class_day_date()}: Format a date as day of week, month, and day.

}}
\examples{
format_class_date(Sys.Date())

format_class_day_date(Sys.Date())

}
