% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_formatting.R
\name{format_date_range}
\alias{format_date_range}
\alias{format_day_date_range}
\title{Format a range of dates}
\usage{
format_date_range(dates, abbr = TRUE)

format_day_date_range(dates, abbr_month = TRUE, abbr_wday = TRUE)
}
\arguments{
\item{dates}{A named list or vector of dates with elements \code{start} and \code{stop}.}

\item{abbr}{Abbreviate the month.}

\item{abbr_month}{Abbreviate the months.}

\item{abbr_wday}{Abbreviate the days of the week.}
}
\value{
A formatted date
}
\description{
Format a range of dates.
}
\details{
Format a range of dates, using a similar format to
\code{\link{format_class_date}}
}
\section{Functions}{
\itemize{
\item \code{format_day_date_range}: Format a range of dates, including day of the week.
}}

\examples{
format_date_range(list(start = lubridate::ymd("2020-02-14"),
                       stop  = lubridate::ymd("2020-03-15")))

format_day_date_range(list(start = lubridate::ymd("2020-02-14"),
                           stop  = lubridate::ymd("2020-03-15")))

}
