% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_formatting.R
\name{format_date_range_by_event_id}
\alias{format_date_range_by_event_id}
\title{FUNCTION_TITLE}
\usage{
format_date_range_by_event_id(
  calendar,
  event_ids,
  abbr = TRUE,
  days = FALSE,
  abbr_wday = NULL
)
}
\arguments{
\item{calendar}{DESCRIPTION.}

\item{event_ids}{DESCRIPTION.}

\item{abbr}{DESCRIPTION.}

\item{days}{DESCRIPTION.}

\item{abbr_wday}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
