% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_formatting.R
\name{format_date_range_by_key}
\alias{format_date_range_by_key}
\title{FUNCTION_TITLE}
\usage{
format_date_range_by_key(
  calendar,
  keys,
  type = c("class", "reading", "lab", "homework", "exam", "holiday", "event",
    "due date", "raw"),
  abbr = TRUE,
  days = FALSE,
  abbr_wday = NULL
)
}
\arguments{
\item{calendar}{DESCRIPTION.}

\item{keys}{DESCRIPTION.}

\item{type}{DESCRIPTION.}

\item{abbr}{DESCRIPTION.}

\item{days}{DESCRIPTION.}

\item{abbr_wday}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
