% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_assignments.R
\name{generate_assignments}
\alias{generate_assignments}
\title{Generate assignments from database}
\usage{
generate_assignments(semester, dry_run = FALSE)
}
\arguments{
\item{semester}{A semester object returned from
\code{\link[=load_semester_db]{load_semester_db()}}.}

\item{dry_run}{Don't actually write assignment files to disk.}
}
\value{
A named list containing the lesson plan in YAML text format and
the semester schedule, as a \code{\link[tibble:tbl_df-class]{tibble}}.
}
\description{
Generate RMarkdown files for reading, homework, and lab  assignments and
a lessons.yml file for Hugo to use in making a schedule for a course.
}
\examples{
\dontrun{
sem <- load_semester_db("foo.sqlite3")
asgts <- generate_assignments(sem)
}
}
