% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_digests.R
\name{get_one_pdf_url}
\alias{get_one_pdf_url}
\title{Get URL corresponding to a generated}
\usage{
get_one_pdf_url(file)
}
\arguments{
\item{file}{The path to the RMarkdown source.}
}
\value{
The URL for the generated PDF file.
}
\description{
From an RMarkdown source file, get the URL for downloading the PDF file that
will be generated from it.
}
\keyword{internal}
