% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_digests.R
\name{get_pdf_digest_algorithm}
\alias{get_pdf_digest_algorithm}
\title{Get the digest algorithm to use}
\usage{
get_pdf_digest_algorithm()
}
\value{
A string containing the name of the algorithm.
}
\description{
\code{get_pdf_digest_algorithm} gets the digest algorithm that will be used.
}
\details{
Set the algorithm with \verb{options(blogdown.hash.algorithm = <algorithm>)}.
If the option is not set, then use crc32.
}
\seealso{
\link[blogdownDigest:digests]{blogdownDigest::digests}.
}
\keyword{internal}
