% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_digests.R
\name{get_pdf_url}
\alias{get_pdf_url}
\title{Get URLs corresponding to generated PDF files.}
\usage{
get_pdf_url(files)
}
\arguments{
\item{files}{A list or character vector of paths tor the RMarkdown
source files.}
}
\value{
A character vector with the URLs for the generated PDF files.
}
\description{
From a collection of RMarkdown source files, get the URLs for downloading
the PDF files that will be generated from them.
}
