% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_utils.R
\name{get_semestr_metadata}
\alias{get_semestr_metadata}
\title{Get the metadata for the currently loaded semester}
\usage{
get_semestr_metadata()
}
\value{
A named list with the metadata
}
\description{
Pulls the metadata for the currently loaded semester from the package's
\code{.globals}  environment
}
\examples{
get_semestr_metadata()

}
