% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_utils.R
\name{get_semestr_tz}
\alias{get_semestr_tz}
\title{Get the time zone for where the course will be taught.}
\usage{
get_semestr_tz()
}
\value{
A character string with the time zone.
}
\description{
The semester planning database should contain information about the
time zone where the course will be taught. This function retrieves the
time zone corresponding to the currently loaded semester schedule.
}
\examples{
get_semestr_tz()

}
