% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_search.R
\name{global_col_search}
\alias{global_col_search}
\title{Search data frames for a pattern in any column}
\usage{
global_col_search(df_lst, pattern)
}
\arguments{
\item{df_lst}{A named list of data frames}

\item{pattern}{The pattern to search for (regular expression)}
}
\value{
A named list of data frames where character columns are replaced
with "HIT" for matches and "" for misses.
}
\description{
Search data frames and return lines where any column matches a pattern.
}
