% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_search.R
\name{global_col_search_sum}
\alias{global_col_search_sum}
\title{Count pattern-matches in each column of data frames}
\usage{
global_col_search_sum(df_lst, pattern)
}
\arguments{
\item{df_lst}{A named list of data frames}

\item{pattern}{The pattern to search for (regular expression)}
}
\value{
A named list of data frames where each column has the number of
times that column matched the pattern.
}
\description{
Search a list of data frames for a pattern and count the number of matches
in each column.
}
