% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_search.R
\name{global_replace}
\alias{global_replace}
\title{Search and replace all columns in a list of data frames.}
\usage{
global_replace(df_lst, pattern, replacement)
}
\arguments{
\item{df_lst}{A named list of data frames}

\item{pattern}{The pattern to search for (regular expression)}

\item{replacement}{The replacement.}
}
\value{
A named list of data frames.
}
\description{
Search and replace in every column of a list of data frames.
}
