% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_header.R
\name{grab_header}
\alias{grab_header}
\title{Extract metadata from an RMD file.}
\usage{
grab_header(f, buflines = 20)
}
\arguments{
\item{f}{The filename}

\item{buflines}{How big a chunk to read in at a time.}
}
\value{
A data frame containing the header metadata.
}
\description{
Extract metadata from the YAML header in an RMD file.
}
