% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_conversions.R
\name{idx2type}
\alias{idx2type}
\title{FUNCTION_TITLE}
\usage{
idx2type(idx)
}
\arguments{
\item{idx}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
