% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_assignments.R
\name{init_schedule}
\alias{init_schedule}
\title{Initialize semester schedule}
\usage{
init_schedule(semester)
}
\arguments{
\item{semester}{A semester object (a list) returned from
\code{\link[=load_semester_db]{load_semester_db()}}}
}
\value{
A data frame containing the semester schedule.
}
\description{
Convert a semester data object into a schedule suitable for
generating assignments.
}
