% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_utils.R
\name{is_mt_or_na}
\alias{is_mt_or_na}
\title{Extends is.na to report TRUE if the object has length zero.}
\usage{
is_mt_or_na(x)
}
\arguments{
\item{x}{A vector of any type, of length 0 or 1}
}
\value{
TRUE if the vector has length 0 or is NA
}
\description{
If I want to check whether an element in a list is missing, sometimes it's
\code{NA}, but sometimes it is simply a vector of length 0. This checks for both
possibilities.
}
\examples{
is_mt_or_na(character(0))
is_mt_or_na(NA_character_)
is_mt_or_na("foo")

}
