% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_semester_db.R
\name{item_type}
\alias{item_type}
\title{Determine the type of calendar entry from its calendar id.}
\usage{
item_type(cal_id, metadata)
}
\arguments{
\item{cal_id}{an integer calendar ID number.}

\item{metadata}{A list of metadata as returned from \link{load_semester_db}.}
}
\value{
A string identifying the type of calendar entry. Current values
are "class", "reading", "homework", "lab", "exam", "due date", "holiday",
and "event".
}
\description{
Determine the type of calendar entry from its calendar id.
}
