% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_semester_db.R
\name{load_semester_db}
\alias{load_semester_db}
\title{Load schedule for semester from database}
\usage{
load_semester_db(db_file, root_crit = NULL, ignore_root = FALSE)
}
\arguments{
\item{db_file}{An SQLite database file.}
}
\value{
A list containing a tibbles with the calendar for the semester and
tibbles with details on reading assignments, homework assignments,
labs, exams, holidays, and other events.

THe list contains the following tibbles:
\code{calendar}, \code{due_dates}, \code{rd_items}, \code{rd_src},
\code{class_topics}, \code{hw_asgt}, \code{hw_items}, \code{hw_sol}, \code{hw_topics},
\code{lab_asgt}, \code{lab_items}, \code{lab_sol}, \code{exams},
\code{holidays}, \code{events}, \code{notices},
\code{text_codes}

and a named list \code{metadata} containing named character vectors of
metadata that are used to decode and manipulate calendar entries.
}
\description{
Loads schedule for class meetings, reading and homework assignments,
lab sessions, exams, holidays, etc. from a database.
}
