% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_formatting.R
\name{lookup_and_format_date}
\alias{lookup_and_format_date}
\alias{format_date_by_cal_id}
\alias{format_date_by_class_num}
\alias{format_date_by_key}
\alias{format_day_date_by_cal_id}
\alias{format_day_date_by_class_num}
\alias{format_day_date_by_key}
\title{Look up and format a date}
\usage{
format_date_by_cal_id(calendar, id, abbr = TRUE)

format_date_by_class_num(calendar, num, abbr = TRUE)

format_date_by_key(
  calendar,
  key,
  type = c("class", "reading", "lab", "homework", "exam", "holiday", "event",
    "due date", "raw"),
  abbr = TRUE
)

format_day_date_by_cal_id(calendar, id, abbr_month = TRUE, abbr_wday = TRUE)

format_day_date_by_class_num(
  calendar,
  num,
  abbr_month = TRUE,
  abbr_wday = TRUE
)

format_day_date_by_key(
  calendar,
  key,
  type = c("class", "reading", "lab", "homework", "exam", "holiday", "event",
    "due date", "raw"),
  abbr_month = TRUE,
  abbr_wday = TRUE
)
}
\arguments{
\item{calendar}{The calendar to use for lookin up the date.}

\item{id}{The \code{cal_id} index to look up.}

\item{abbr}{Abbreviate the month.}

\item{num}{The class session number to look up.}

\item{key}{The value to look up.}

\item{type}{Which index to use in the lookup ("class", "reading", "lab",
"Homework", "exam", "holiday", "event", "due date", or "raw"..
'}

\item{abbr_month}{Abbreviate the month.}

\item{abbr_wday}{Abbreviate the day of the week.}
}
\description{
Look up a date from the semester schedule and format it.
}
\details{
If the key type is \code{"raw"}, then the fully formatted key should
be used (e.g., "CLS_INTRO" for a class key). Otherwise, the prefix will
be added automatically based on the type, so \verb{key = "INTRO", type = "lab"}
will become \code{LAB_INTRO}.
}
\section{Functions}{
\itemize{
\item \code{format_date_by_cal_id}: Look up a date by calendar id

\item \code{format_date_by_class_num}: Look up a date by class session

\item \code{format_date_by_key}: Look up a date by arbitrary key.

\item \code{format_day_date_by_cal_id}: Look up a date by calendar id.

\item \code{format_day_date_by_class_num}: Look up a date by class session.

\item \code{format_day_date_by_key}: Look up a date by arbitrary key.
}}

\examples{
\dontrun{
format_date_by_key(calendar, "IR_RADIATION", "class")
}

\dontrun{
format_date_by_key(calendar, "IR_RADIATION", "class")
}

\dontrun{
format_day_date_by_class_num(calendar, 11)
}
}
