% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_formatting.R
\name{lookup_and_format_date_range}
\alias{lookup_and_format_date_range}
\alias{format_date_range_by_cal_id}
\alias{format_date_range_by_class_num}
\alias{format_date_range_by_key}
\alias{format_date_range_by_event_id}
\title{Look up and format a range of dates}
\usage{
format_date_range_by_cal_id(
  calendar,
  cal_ids,
  abbr = TRUE,
  days = FALSE,
  abbr_wday = NULL
)

format_date_range_by_class_num(
  calendar,
  nums,
  abbr = TRUE,
  days = FALSE,
  abbr_wday = NULL
)

format_date_range_by_key(
  calendar,
  keys,
  type = c("class", "reading", "lab", "homework", "exam", "holiday", "event", "due date",
    "raw"),
  abbr = TRUE,
  days = FALSE,
  abbr_wday = NULL
)

format_date_range_by_event_id(
  calendar,
  event_ids,
  abbr = TRUE,
  days = FALSE,
  abbr_wday = NULL
)
}
\arguments{
\item{calendar}{The calendar to use for looking up the date.}

\item{cal_ids}{A list or vector of calendar IDs.}

\item{abbr}{Abbreviate the month.}

\item{days}{Include day of week in formatted date.}

\item{abbr_wday}{Abbreviate weekdays (if this is \code{NULL}, use \code{abbr}).}

\item{nums}{A list or vector of class numbers.}

\item{keys}{A list or vector of key indices to look up.}

\item{type}{Which type of key is this ("class", "reading", "lab",
"homework", "exam", "holiday", "event", "due date", or "raw")}

\item{event_ids}{A list or vector of event ids.}
}
\value{
A formatted date range
}
\description{
Look up dates from the semester schedule and format them as a range from
earliest to latest.
}
\details{
If the key type is \code{"raw"}, then the fully formatted key should
be used (e.g., "CLS_INTRO" for a class key). Otherwise, the prefix will
be added automatically based on the type, so \verb{key = "INTRO", type = "lab"}
will become \code{LAB_INTRO}.
}
\section{Functions}{
\itemize{
\item \code{format_date_range_by_cal_id()}: Look up dates by calendar ID.

\item \code{format_date_range_by_class_num()}: Look up dates by class number.

\item \code{format_date_range_by_key()}: Look up dates by key

\item \code{format_date_range_by_event_id()}: Look up dates by event id.

}}
