% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_referencing.R
\name{lookup_class}
\alias{lookup_class}
\title{Look Up Classes}
\usage{
lookup_class(calendar, key)
}
\arguments{
\item{calendar}{The calendar to use for looking up the classes.}

\item{key}{The \code{key_class} index of the target class.}
}
\value{
A row from the calendar table.
}
\description{
Look up classes from the class key.
}
