% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_reading.R
\name{make_reading_assignment}
\alias{make_reading_assignment}
\title{Make a Reading Assignment}
\usage{
make_reading_assignment(reading_entry)
}
\arguments{
\item{reading_entry}{A reading entry assembled from the semester database
by \verb{\link{make_reading_page}}}
}
\value{
A formatted reading assignment, suitable for including in an
assignment page.
}
\description{
Format all readings from a reading assignment entry.
}
\keyword{internal}
