% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_reading.R
\name{make_reading_page}
\alias{make_reading_page}
\title{Make a Reading Assignment Page}
\usage{
make_reading_page(cal_id, semester)
}
\arguments{
\item{cal_id}{The calendar ID for this assignment.}

\item{semester}{A semester object (a list returned by
\verb{\link{load_semester_db}}).}
}
\value{
A character string with the Markdown content for the reading
assignment page.
}
\description{
Make a page for the reading assignment for one class, from the semester
database.
}
