% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_digests.R
\name{pdf_digest_if_exists}
\alias{pdf_digest_if_exists}
\title{Calculate digest for pdf file}
\usage{
pdf_digest_if_exists(file, alg = NA_character_)
}
\arguments{
\item{file}{The pdf file (the file need not exist).}

\item{alg}{The digest algorithm to use.}
}
\value{
A character vector with the digest, or \code{NA} if the file does not
exist.
}
\description{
Check whether a PDF file exists, and if so calculate its digest.
}
