% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pdf_files.R
\name{pdf_filename}
\alias{pdf_filename}
\title{Extract the local PDF filename corresponding to an URL.}
\usage{
pdf_filename(
  pdf_url,
  root_dir,
  static_path = "static",
  force_dest = FALSE,
  ignore_missing_dest = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{pdf_url}{The URL (in the destination web site).}

\item{root_dir}{The root directory of the local HUGO project.}

\item{static_path}{Relative path to the HUGO static files.}

\item{force_dest}{Create missing directories?}

\item{ignore_missing_dest}{Ignore missing directories.}

\item{verbose}{Provide informative messages and warnings.}
}
\value{
The filename
}
\description{
Extract the path to a local PDF file corresponding to an URL on the web site.
}
