% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_digests.R
\name{pdfs_to_rebuild}
\alias{pdfs_to_rebuild}
\title{Figure out which files need to be rebuilt}
\usage{
pdfs_to_rebuild(
  files,
  root_dir,
  static_path = "static",
  content_path = "content"
)
}
\arguments{
\item{files}{A character vector of paths to source files (e.g., \code{.Rmd}).}
}
\value{
A character vector of files that need to be rebuilt.
}
\description{
\code{pdfs_to_rebuild} returns a vector of files that need to be rebuilt.
}
\details{
This function accepts a vector of source files and
returns a vector of files that need to be rebuilt because the source file is
new or has changed since the last time the site was built.
}
\seealso{
\code{\link[=get_current_pdf_digests]{get_current_pdf_digests()}},
\link[blogdownDigest:digests]{blogdownDigest::digests}.
}
\keyword{internal}
