% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_assignments.R
\name{prepare_schedule}
\alias{prepare_schedule}
\title{Prepare schedule from database}
\usage{
prepare_schedule(semester)
}
\arguments{
\item{semester}{A semester object returned from
\code{\link{load_semester_db}}.}
}
\value{
A tibble containing a schedule
}
\description{
Prepare schedule from database
}
\examples{
\dontrun{
sem <- load_semester_db("foo.sqlite3")
sched <- prepare_schedule(sem)
}
}
