% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_digests.R
\name{prune_pdf_digests}
\alias{prune_pdf_digests}
\title{Delete stored digests for specified source files}
\usage{
prune_pdf_digests(files, root_dir = NULL)
}
\arguments{
\item{files}{A character vector of paths to the source files to be removed.}

\item{root_dir}{The root directory of the HUGO project.}
}
\value{
The path to the digest file.
}
\description{
\code{prune_pdf_digests} removes the lines from the digest file
corresponding to a vector of source files.
}
\details{
Modifies the stored digest file to remove lines corresponding to selected
source files.
}
\seealso{
\code{\link{update_site_digests}()}, \code{\link{digests}}.
}
