% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_database.R
\name{reindex}
\alias{reindex}
\title{Re-index calendar entries of a single type.}
\usage{
reindex(df, type, key_col = NULL, idx_col = NULL)
}
\arguments{
\item{df}{A dataframe holding calendar events of a given type.
Columns should include:
\describe{
\item{cal_id}{An integer whose thousands place indicates the kind of
entry, hundreds place indicates re-scheduling (cancellation or make-up),
and last two digits indicate the order of the entry.}
\item{date}{The date of the entry, as an ISO format string.}
\item{cancelled}{An integer indicating whether the entry has been
cancelled.}
\item{make_up}{An integer indicating whether the entry represents a
re-scheduled make-up session.}
\item{\code{key_col}}{A column with a text key for that entry.}
\item{\code{idx_col}}{A column with an integer index representing the sequence
of this kind of entry (class, lab, etc.)}
\item{cal_type}{A string indicating the type of entry (class, lab, etc.)}
}}

\item{key_col}{The name of the column with the keys for the entries.
This can either be quoted or unquoted (i.e., using non-standard tidy
evaluation).}

\item{idx_col}{The name of the column with the integer index for entries
\code{homework}, \verb{due date}, \code{exam}, \code{holiday}, or \code{event}.
If \code{idx_col} is \code{NULL}, the value is constructed by replacing "\verb{_key}" with
"\verb{_num}" in the \code{key_col} argument (e.g., "\code{class_key}" becomes
"\code{class_num}").}
}
\value{
A tibble with the same columns as the input \code{df}, but with the
numerical index in \code{idx_col} adjusted so multiple entries with the same
key in \code{key_col} have the same numerical index in \code{idx_col}.
}
\description{
Take a calendar with entries of a given type and re-index them so that
repeated entries with the same text key get the same numerical index.
For instance, when there is a week-long holiday, such as Thanksgiving or
Spring Break, there may be two or three entries with the same key, but on
different dates. This function makes sure that they are all assigned the
same numerical index.
}
