% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_assignments.R
\name{schedule_add_reading}
\alias{schedule_add_reading}
\title{Add reading assignments to a schedule}
\usage{
schedule_add_reading(schedule, semester)
}
\arguments{
\item{schedule}{A \code{schedule} data frame, as returned from
\code{\link[=init_schedule]{init_schedule()}}.}

\item{semester}{A \code{semester} object (a list), as returned from
\code{\link[=load_semester_db]{load_semester_db()}}}
}
\value{
A \code{schedule} data frame, with the reading assignments
added.
}
\description{
Add reading assignments to a \code{schedule} data frame.
}
