% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_assignments.R
\name{schedule_strip_finals}
\alias{schedule_strip_finals}
\title{Strip the final exams from a schedule}
\usage{
schedule_strip_finals(schedule, semester)
}
\arguments{
\item{schedule}{A schedule data frame, as returned from
\code{\link[=init_schedule]{init_schedule()}}.}

\item{semester}{A semester object (a list), returned from
\code{\link[=load_semester_db]{load_semester_db()}}.}
}
\value{
A list of a \code{schedule} data frame with the exams removed,
and a \code{final_examss} data frame containing the final exams rows.
}
\description{
Remove the final exams from a semester schedule.
}
