% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_assignments.R
\name{schedule_widen}
\alias{schedule_widen}
\title{Pivot a schedule data frame into wide format.}
\usage{
schedule_widen(
  schedule,
  final_exams,
  semester,
  final_is_take_home = TRUE,
  create_paths = TRUE
)
}
\arguments{
\item{schedule}{A \code{schedule} data frame, as returned from
\code{\link[=init_schedule]{init_schedule()}}.}

\item{final_exams}{DESCRIPTION.}

\item{semester}{A \code{semester} object (a list), as returned from
\code{\link[=load_semester_db]{load_semester_db()}}}

\item{final_is_take_home}{DESCRIPTION.}

\item{create_paths}{DESCRIPTION.}
}
\value{
A \code{schedule} data frame, converted to wide format.
}
\description{
Combine a \code{schedule} data frame, containing class descriptions,
reading assignments, and homework assignments, along with
other data, merge them, and pivot to a wide format, where
every calendar entry has a column for the reading assignment,
homework due that day, and other information.
}
