% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_prefixes.R
\name{strip_key_prefix.data.frame}
\alias{strip_key_prefix.data.frame}
\title{Strip prefixes off keys.}
\usage{
\method{strip_key_prefix}{data.frame}(df, type, col = "cal_key", ...)
}
\arguments{
\item{df}{The data frame to process.}

\item{type}{The type of calendar entry to strip (e.g., "\code{class}", "\code{lab}",
etc.)}

\item{col}{The column where the keys are located (by default "\code{cal_key}").}
}
\value{
A data frame with the prefixes stripped from the specified column.
}
\description{
Keys in the master calendar's \code{cal_key} column have prefixes according to
the type of calennder entry they represent (e.g., "\code{LAB_}" for labs,
"\code{CLS_}" for classes/reading assignments, "\code{EXAM_}" for exams, etc.). This
function strips those off.
}
\seealso{
add_key_prefix
}
