% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_conversions.R
\name{type2col}
\alias{type2col}
\title{FUNCTION_TITLE}
\usage{
type2col(type)
}
\arguments{
\item{type}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
