% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_latex_includes.R
\name{update_latex_styles}
\alias{update_latex_styles}
\title{Update LaTeX style files}
\usage{
update_latex_styles(
  root_dir = NULL,
  content_path = "content",
  planning_path = "planning"
)
}
\arguments{
\item{root_dir}{The root directory for the project (Should be the root for
a HUGO project).}

\item{content_path}{Where to look for content (e.g., \code{.Rmd} and \code{.md} files).}

\item{planning_path}{Where to look for planning files (typically, in the
same directory where the semester database is stored.).}
}
\description{
Update LaTeX style files in content folders, copying from planning folder.
}
