% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_digests.R
\name{update_pdf_digests}
\alias{update_pdf_digests}
\title{Generates and stores digests for all source and output files.}
\usage{
update_pdf_digests(
  dir = NULL,
  root_dir = NULL,
  static_path = "static",
  content_path = "content",
  partial = FALSE
)
}
\arguments{
\item{dir}{A string with the name of the directory to search
(by default the "content" directory at the top-level directory of the site)
If the \code{root_dir} is \code{NULL}, the function tries to find the root directory
using a few common heuristics.}

\item{root_dir}{The root directory for the project (Should be the root for
a HUGO project).}

\item{static_path}{Where to look for static files (in the HUGO sense of
static).}

\item{content_path}{Where to look for content (e.g., \code{.Rmd} and \code{.md} files).}

\item{partial}{Logical. If \code{TRUE}, keep digests for source
files that aren't in the specified directory and its children and
descendants.
Otherwise, get rid of the old digest file and only keep digests for
source files in the source directory and its descendants.}
}
\value{
The path to the digest file.
}
\description{
\code{blogdownDigest::update_site_digests} calculates hashed digests for a
site.
}
\details{
Generates new hashed digests for both source and destination (output) files
and save the digests to a file "\code{digests.Rds}" in the root directory of
the site.
}
\seealso{
\code{\link[=prune_pdf_digests]{prune_pdf_digests()}},
\code{\link[blogdownDigest]{update_site}()},
\link[blogdownDigest:digests]{blogdownDigest::digests}.
}
