% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_digests.R
\name{update_pdf_dir}
\alias{update_pdf_dir}
\title{Rebuild changed files in a subdirectory of "content"}
\usage{
update_pdf_dir(
  dir = ".",
  root_dir = NULL,
  static_path = "static",
  content_path = "content",
  quiet = TRUE,
  force = FALSE,
  ignore = NA
)
}
\arguments{
\item{dir}{A string containing the root directory for checking.
By default, the "content" directory of the project.}

\item{root_dir}{The root directory for the project (Should be the root for
a HUGO project).}

\item{static_path}{Where to look for static files (in the HUGO sense of
static).}

\item{content_path}{Where to look for content (e.g., \code{.Rmd} and \code{.md} files).}

\item{quiet}{Suppress output. By default this is \code{FALSE} and the
function emits an informational message about how many files will
be rebuilt.}

\item{force}{Force rebuilding source files that are not out of date.}

\item{ignore}{A regular expression pattern for files to ignore.}
}
\value{
A list of out-of-date files to rebuild.
}
\description{
\code{update_dir} updates changed files in a subdirectory of "content"
}
