% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_digests.R
\name{update_pdf_file_digests}
\alias{update_pdf_file_digests}
\title{Update digests for PDF files}
\usage{
update_pdf_file_digests(
  files,
  root_dir,
  static_path = "static",
  content_path = "content",
  partial = FALSE
)
}
\arguments{
\item{files}{Files to update.}

\item{root_dir}{The root directory for the project (Should be the root for
a HUGO project).}

\item{static_path}{Where to look for static files (in the HUGO sense of
static).}

\item{content_path}{Where to look for content (e.g., \code{.Rmd} and \code{.md} files).}

\item{partial}{Only update digests for new files without previous digests..}
}
\value{
A data frame with the digests.
}
\description{
Update digests for all PDF files under the content path.
}
\details{
If the \code{root_dir} is \code{NULL}, the function tries to find the root directory
using a few common heuristics.
}
