% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_semester_db.R
\name{add_key_prefix}
\alias{add_key_prefix}
\title{Add prefixes to keys.}
\usage{
add_key_prefix(df, type, key_col = "topic_key")
}
\arguments{
\item{df}{The data frame to process.}

\item{type}{The type of prefix to add (e.g., "\code{class}", "\code{lab}",
etc.)}

\item{metadata}{A list containing metadata for the differennt types.}

\item{col}{The column where the keys are located (by default "\code{topic_key}").}
}
\value{
A data frame with the prefixes stripped from the specified column.
}
\description{
Keys in the master calendar's \code{topic_key} column have prefixes according to
the type of calennder entry they represent (e.g., "\code{LAB_}" for labs,
"\code{CLS_}" for classes/reading assignments, "\code{EXAM_}" for exams, etc.). This
function adds those prefixes.
}
\seealso{
strip_key_prefix
}
