% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_assignments.R
\name{build_assignments}
\alias{build_assignments}
\title{Build reading, homework, and lab assignments}
\usage{
build_assignments(schedule, semester)
}
\arguments{
\item{schedule}{A schedule dataframe.}

\item{semester}{A semester object (list).}
}
\value{
An updated schedule dataframe
}
\description{
Build reading, homework, and lab assignments from a
schedule dataframe
}
