% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_assignments.R
\name{build_lab_assignment}
\alias{build_lab_assignment}
\title{Build a lab assignment}
\usage{
build_lab_assignment(schedule, date, cal_entry, semester)
}
\arguments{
\item{schedule}{A schedule data frame.}

\item{date}{The date of the lab.}

\item{cal_entry}{A calendar entry for the lab session.}

\item{semester}{A list of data for the semester, from the database.}
}
\value{
An updated schedule data frame
}
\description{
Build a lab assignment: Generate an \code{.Rmd} file for a lab assignment
web page and PDF handout.
}
