% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pdf_files.R
\name{build_pdf_from_rmd}
\alias{build_pdf_from_rmd}
\title{Build a PDF from an RMarkdown source file}
\usage{
build_pdf_from_rmd(
  source_file,
  root_dir,
  static_path = "static",
  force_dest = FALSE
)
}
\arguments{
\item{source_file}{The RMarkdown source file}

\item{root_dir}{The root directory of the HUGO project.}

\item{static_path}{Relative path to the HUGO static directory.}

\item{force_dest}{Create any missing directories.}
}
\value{
The path to the resulting PDF file.
}
\description{
Build a PDF from an RMarkdown source file
}
