% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pdf_files.R
\name{build_pdf_output_format}
\alias{build_pdf_output_format}
\title{Build an RMarkdown output format for PDF files.}
\usage{
build_pdf_output_format(header)
}
\arguments{
\item{header}{Metadata from the RMarkdown file.}
}
\value{
An RMarkdown output format object.
}
\description{
Build an RMarkdown output format for PDF files.
}
