% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_assignments.R
\name{copy_slides}
\alias{copy_slides}
\title{Install Powerpoint slides for a class session}
\usage{
copy_slides(schedule, date, cal_entry, semester)
}
\arguments{
\item{schedule}{A schedule data frame.}

\item{date}{The reading due date.}

\item{cal_entry}{A calendar entry for the class.}

\item{semester}{A list of data for the semester, from the database.}
}
\value{
An updated schedule data frame
}
\description{
For class sessions where a Powerpoint slide deck will be used
instead of \code{reveal.js} slides, copy the Powerpoint slides from
a main directory to the class slide directory.
}
\details{
Normally, \code{semestr} assumes that the user will be using \code{reveal.js},
and that the slides will be installed (e.g., by using \verb{gulp sync}
from the directory where the slides are authored), but sometimes,
especially for guest lectures, I use Powerpoint, and this lets me
copy files from a main directory where I save Powerpoint slides into
the \verb{/slides/} directory for the website.
}
