% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_utils.R
\name{default_semestr_metadata}
\alias{default_semestr_metadata}
\title{Create default metadata}
\usage{
default_semestr_metadata()
}
\value{
A named list whose elements are named vectors corresponding to the
type of syllabus entries, the data frame columns used to store data for
those events, prefixes used in calendar keys, and numeric bases used for
calendar ID entries, as well as modifiers for calendar IDs used to
indicate canceled and rescheduled classes.
}
\description{
\code{default_semestr_metadata} creates a named list with metadata for the
structure of a syllabus.
}
