% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_formatting.R
\name{format_page_range}
\alias{format_page_range}
\title{FUNCTION_TITLE}
\usage{
format_page_range(pages)
}
\arguments{
\item{pages}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
