% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_search.R
\name{global_search}
\alias{global_search}
\title{Search data frames for a pattern in any column.}
\usage{
global_search(df_lst, pattern)
}
\arguments{
\item{pattern}{The pattern to search for (regular expression)}

\item{df_list}{A named list of data frames}
}
\value{
A named list of filtered data frames
}
\description{
Search data frames and return lines where any column matches a pattern.
}
