% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_assignments.R
\name{init_schedule}
\alias{init_schedule}
\title{FUNCTION_TITLE (TODO)}
\usage{
init_schedule(semester)
}
\arguments{
\item{semester}{DESCRIPTION  (TODO).}
}
\value{
RETURN_DESCRIPTION  (TODO)
}
\description{
FUNCTION_DESCRIPTION  (TODO)
}
\examples{
# ADD_EXAMPLES_HERE  (TODO)
}
