% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_conversions.R
\name{item_mod}
\alias{item_mod}
\title{Determine the modification type of calendar entry from its calendar id.}
\usage{
item_mod(cal_id)
}
\arguments{
\item{cal_id}{an integer calendar ID number.}
}
\value{
A string identifying the type of modification. Current values are
"cancelled" and "make-up"
}
\description{
Modifications include cancelled and re-scheduled (make-up) classes.
}
