% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_referencing.R
\name{lookup_past_future_classes}
\alias{lookup_past_future_classes}
\alias{lookup_future_class}
\alias{lookup_past_class}
\title{Look Up Past and Future Classes}
\usage{
lookup_future_class(calendar, id, delta = 1)

lookup_past_class(calendar, id, delta = 1)
}
\arguments{
\item{calendar}{The calendar to use for looking up the classes.}

\item{id}{The \code{cal_id} index of the current class.}

\item{delta}{The number of classes forward or backward.}
}
\value{
A row from the calendar table.
}
\description{
Look up previous or future classes, based on the current class's id.
}
\section{Functions}{
\itemize{
\item \code{lookup_future_class()}: 

\item \code{lookup_past_class()}: 

}}
