% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_reading.R
\name{make_reading_page}
\alias{make_reading_page}
\title{Make a Reading Assignment Page}
\usage{
make_reading_page(cal_id, semester, schedule, use_pdfs = TRUE)
}
\arguments{
\item{cal_id}{The calendar ID for this assignment.}

\item{semester}{A semester object (a list returned by
\verb{\link{load_semester_db}}).}

\item{schedule}{A data frame with the semester schedule (returned by
\link{prepare_schedule})}

\item{use_pdfs}{Add a \code{pdf_url} field to the YAML header so a PDF file will
be generated for the reading assignment.}
}
\value{
A character string with the Markdown content for the reading
assignment page.
}
\description{
Make a page for the reading assignment for one class, from the semester
database.
}
