% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_output_format.R
\name{make_rmd_output_format}
\alias{make_rmd_output_format}
\title{Make an RMD output format}
\usage{
make_rmd_output_format(toc = FALSE, toc_depth = NULL, includes = NULL)
}
\arguments{
\item{toc}{Include a table of contents?}

\item{toc_depth}{Depth of the table of contents}

\item{includes}{list of files to include.}
}
\value{
A list containing a named list of output formats.
}
\description{
Make an output format for creating blogdown and PDF output.
}
