% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_semester_db.R
\name{pull_env}
\alias{pull_env}
\title{Expose contents of an environment in the current environment}
\usage{
pull_env(env)
}
\arguments{
\item{env}{The environment to attach locally.}
}
\description{
Expose the contents of an environment in the current environment.
Similar to \link{attach}, but it exposes the contents only in the
current environment and does not change the global search.
}
