% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_assignments.R
\name{schedule_add_reading}
\alias{schedule_add_reading}
\title{FUNCTION_TITLE}
\usage{
schedule_add_reading(schedule, semester)
}
\arguments{
\item{schedule}{DESCRIPTION.}

\item{semester}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
