% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_assignments.R
\name{schedule_strip_finals}
\alias{schedule_strip_finals}
\title{FUNCTION_TITLE  (TODO)}
\usage{
schedule_strip_finals(schedule, semester)
}
\arguments{
\item{schedule}{DESCRIPTION  (TODO).}

\item{semester}{DESCRIPTION  (TODO).}
}
\value{
RETURN_DESCRIPTION  (TODO)
}
\description{
FUNCTION_DESCRIPTION  (TODO)
}
\examples{
# ADD_EXAMPLES_HERE  (TODO)
}
