% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_assignments.R
\name{schedule_widen}
\alias{schedule_widen}
\title{FUNCTION_TITLE}
\usage{
schedule_widen(
  schedule,
  final_exams,
  semester,
  final_is_take_home = TRUE,
  create_paths = TRUE
)
}
\arguments{
\item{schedule}{DESCRIPTION.}

\item{final_exams}{DESCRIPTION.}

\item{semester}{DESCRIPTION.}

\item{final_is_take_home}{DESCRIPTION.}

\item{create_paths}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
