% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semestr-package.R
\docType{package}
\name{semestr-package}
\alias{semestr}
\alias{semestr-package}
\title{semestr: Generate Assignments for a Semester from a SQLlite Database}
\description{
Manage a syllabus, reading, homework, and lab assignments, and a course web site from a SQL database.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://j-magnolia.github.io/semestr/}
  \item \url{https://github.com/j-magnolia/semestr}
  \item Report bugs at \url{https://github.com/j-magnolia/semestr/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Gilligan \email{j.magnolia.gilligan@vanderbilt.edu} (\href{https://orcid.org/0000-0003-1375-6686}{ORCID})

Authors:
\itemize{
  \item Jonathan Gilligan \email{j.magnolia.gilligan@vanderbilt.edu} (\href{https://orcid.org/0000-0003-1375-6686}{ORCID})
}

}
\keyword{internal}
