% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semestr-package.R
\docType{package}
\name{semestr-package}
\alias{semestr}
\alias{semestr-package}
\title{semestr: Generate assignments for a semester from a SQLlite database}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Manage a syllabus, reading, homework, and lab assignments, and a course
    web site from a SQL database.
}
\author{
\strong{Maintainer}: Jonathan Gilligan \email{jonathan.gilligan@vanderbilt.edu} (\href{https://orcid.org/0000-0003-1375-6686}{ORCID})

}
\keyword{internal}
