% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_prefixes.R
\name{strip_key_prefix}
\alias{strip_key_prefix}
\title{Strip prefixes off keys.}
\usage{
strip_key_prefix(x, type, ...)
}
\arguments{
\item{x}{The list or vector to process.}

\item{type}{The type of calendar entry to strip (e.g., "\code{class}", "\code{lab}",
etc.)}
}
\value{
A list or vector frame with the prefixes stripped from the contents.
}
\description{
Keys in the master calendar's \code{cal_key} column have prefixes according to
the type of calennder entry they represent (e.g., "\code{LAB_}" for labs,
"\code{CLS_}" for classes/reading assignments, "\code{EXAM_}" for exams, etc.). This
function strips those off.
}
\seealso{
add_key_prefix
}
