% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_digests.R
\name{update_pdfs}
\alias{update_pdfs}
\title{Update all files that are out of date}
\usage{
update_pdfs(
  dir = NULL,
  root_dir = NULL,
  static_path = "static",
  content_path = "content",
  quiet = FALSE,
  force = FALSE
)
}
\arguments{
\item{dir}{A string containing the root directory for checking.
By default, the "content" directory of the project.}

\item{quiet}{Suppress output. By default this is \code{FALSE} and the
function emits an informational message about how many files will
be rebuilt.}

\item{force}{Force rebuilding source files that are not out of date.}

\item{method}{Different methods to build a website (each with pros and cons).
See \code{\link[blogdown]{build_site}()} for details.}
}
\value{
This function does not return anything
}
\description{
\code{update_pdfs} rebuilds all source files that are new or have changed
since the last time the site was built.
}
\details{
Given a source directory (by default the "content" directory in the
root directory of the project), find all source files (\code{.Rmd} and
\code{.rmarkdown}) in the directory tree under the source directory,
calculate hashed digests of the files, and compare them to a
stored list of digests from the last time the site was built.

If the digests of either the source or output files don't match,
if a source file is new since the last time the site was built,
or if the output file does not exist,
then render the source file.

After rendering any out-of-date files, regenerate the digest list
and saves it to a file.
}
\seealso{
\code{\link[blogdown]{build_site}()}, \code{\link[blogdown]{build_dir}()},
\code{\link{digests}}.
}
